Enforce same linux only compilation style for memory monitor and killing policy as cgroup
