+++ src/ray/raylet/BUILD.bazel
@@ -1,5 +1,10 @@
 load("//bazel:ray.bzl", "ray_cc_binary", "ray_cc_library")
 
+config_setting(
+    name = "is_linux",
+    constraint_values = ["@platforms//os:linux"],
+)
+
 ray_cc_library(
     name = "agent_manager",
     srcs = ["agent_manager.cc"],
@@ -82,7 +87,10 @@ ray_cc_library(
 ray_cc_library(
     name = "worker_interface",
     hdrs = ["worker_interface.h"],
-    visibility = [":__subpackages__"],
+    visibility = [
+        ":__subpackages__",
+        "//src/ray/common:__pkg__",
+    ],
     deps = [
         "//src/ray/common:id",
         "//src/ray/util:process",
@@ -199,14 +207,38 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "worker_killing_policy",
+    name = "worker_killing_policy_factory",
+    srcs = select({
+        ":is_linux": [
+            "worker_killing_policy_group_by_owner_factory.cc",
+        ],
+        "//conditions:default": [
+            "noop_worker_killing_policy_factory.cc",
+        ],
+    }),
+    hdrs = [
+        "worker_killing_policy_factory.h",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":worker_killing_policy_interface",
+    ] + select({
+        ":is_linux": [
+            ":worker_killing_policy_group_by_owner",
+        ],
+        "//conditions:default": [
+            ":noop_worker_killing_policy",
+        ],
+    }),
+)
+
+ray_cc_library(
+    name = "worker_killing_policy_interface",
     srcs = [
         "worker_killing_policy.cc",
-        "worker_killing_policy_group_by_owner.cc",
     ],
     hdrs = [
         "worker_killing_policy.h",
-        "worker_killing_policy_group_by_owner.h",
     ],
     visibility = [":__subpackages__"],
     deps = [
@@ -215,6 +247,37 @@ ray_cc_library(
         "//src/ray/common:lease",
         "//src/ray/common:memory_monitor",
         "@com_google_absl//absl/time",
+    ],
+)
+
+ray_cc_library(
+    name = "worker_killing_policy_group_by_owner",
+    srcs = [
+        "worker_killing_policy_group_by_owner.cc",
+    ],
+    hdrs = [
+        "worker_killing_policy_group_by_owner.h",
+    ],
+    visibility = [":__subpackages__"],
+    deps = [
+        ":worker_interface",
+        ":worker_killing_policy_interface",
+        ":worker_pool",
+        "//src/ray/common:lease",
+        "//src/ray/common:memory_monitor",
+        "@com_google_absl//absl/time",
+        "@com_google_googletest//:gtest_prod",
+    ],
+)
+
+ray_cc_library(
+    name = "noop_worker_killing_policy",
+    hdrs = [
+        "noop_worker_killing_policy.h",
+    ],
+    visibility = [":__subpackages__"],
+    deps = [
+        ":worker_killing_policy_interface",
     ],
 )
 
@@ -243,12 +306,14 @@ ray_cc_library(
         ":runtime_env_agent_client",
         ":wait_manager",
         ":worker",
-        ":worker_killing_policy",
+        ":worker_killing_policy_factory",
+        ":worker_killing_policy_interface",
         ":worker_pool",
         "//src/ray/common:buffer",
         "//src/ray/common:flatbuf_utils",
         "//src/ray/common:lease",
         "//src/ray/common:memory_monitor",
+        "//src/ray/common:memory_monitor_factory",
         "//src/ray/common/cgroup2:cgroup_manager_interface",
         "//src/ray/core_worker:experimental_mutable_object_provider",
         "//src/ray/core_worker_rpc_client:core_worker_client_pool",
