+++ src/ray/raylet/node_manager.h
@@ -734,7 +734,7 @@ class NodeManager : public rpc::NodeManagerServiceHandler,
   void TriggerLocalOrGlobalGCIfNeeded();
 
   /// Creates the callback used in the memory monitor.
-  MemoryUsageRefreshCallback CreateMemoryUsageRefreshCallback();
+  KillWorkersCallback CreateKillWorkersCallback();
 
   /// Creates the detail message for the worker that is killed due to memory running low.
   std::string CreateOomKillMessageDetails(const std::shared_ptr<WorkerInterface> &worker,
@@ -864,9 +864,6 @@ class NodeManager : public rpc::NodeManagerServiceHandler,
   /// Throttler for global gc
   Throttler global_gc_throttler_;
 
-  /// Target being evicted or null if no target
-  std::shared_ptr<WorkerInterface> high_memory_eviction_target_;
-
   /// These classes make up the new scheduler. ClusterResourceScheduler is
   /// responsible for maintaining a view of the cluster state w.r.t resource
   /// usage. ClusterLeaseManager is responsible for queuing, spilling back, and
@@ -915,7 +912,7 @@ class NodeManager : public rpc::NodeManagerServiceHandler,
   int64_t gc_command_sync_version_ = 0;
 
   /// The Policy for selecting the worker to kill when the node runs out of memory.
-  std::shared_ptr<WorkerKillingPolicy> worker_killing_policy_;
+  std::unique_ptr<WorkerKillingPolicy> worker_killing_policy_;
 
   /// Monitors and reports node memory usage and whether it is above threshold.
   std::unique_ptr<MemoryMonitor> memory_monitor_;
