+++ src/ray/raylet/tests/BUILD.bazel
@@ -133,16 +133,17 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "worker_killing_policy_test",
+    name = "worker_killing_policy_group_by_owner_test",
     size = "small",
     srcs = [
-        "worker_killing_policy_test.cc",
+        "worker_killing_policy_group_by_owner_test.cc",
     ],
     tags = ["team:core"],
     deps = [
         ":util",
         "//src/ray/common:lease",
-        "//src/ray/raylet:worker_killing_policy",
+        "//src/ray/raylet:worker_killing_policy_group_by_owner",
+        "//src/ray/raylet:worker_killing_policy_interface",
         "@com_google_googletest//:gtest_main",
     ],
 )
