+++ src/ray/raylet/tests/worker_killing_policy_group_by_owner_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ray/raylet/worker_killing_policy.h"
+#include "ray/raylet/worker_killing_policy_group_by_owner.h"
 
 #include <memory>
 #include <string>
@@ -22,7 +22,7 @@
 #include "gtest/gtest.h"
 #include "ray/common/lease/lease_spec.h"
 #include "ray/raylet/tests/util.h"
-#include "ray/raylet/worker_killing_policy_group_by_owner.h"
+#include "ray/raylet/worker_killing_policy.h"
 
 namespace ray {
 
@@ -73,9 +73,8 @@ class WorkerKillingGroupByOwnerTest : public ::testing::Test {
 TEST_F(WorkerKillingGroupByOwnerTest, TestEmptyWorkerPoolSelectsNullWorker) {
   std::vector<std::shared_ptr<WorkerInterface>> workers;
   auto worker_to_kill_and_should_retry_ =
-      worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
-  auto worker_to_kill = worker_to_kill_and_should_retry_.first;
-  ASSERT_TRUE(worker_to_kill == nullptr);
+      worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot());
+  ASSERT_TRUE(worker_to_kill_and_should_retry_.empty());
 }
 
 TEST_F(WorkerKillingGroupByOwnerTest, TestLastWorkerInGroupShouldNotRetry) {
@@ -96,7 +95,7 @@ TEST_F(WorkerKillingGroupByOwnerTest, TestLastWorkerInGroupShouldNotRetry) {
 
   for (const auto &entry : expected) {
     auto worker_to_kill_and_should_retry_ =
-        worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
+        worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot()).front();
     auto worker_to_kill = worker_to_kill_and_should_retry_.first;
     bool retry = worker_to_kill_and_should_retry_.second;
     ASSERT_EQ(worker_to_kill->WorkerId(), entry.first->WorkerId());
@@ -121,7 +120,7 @@ TEST_F(WorkerKillingGroupByOwnerTest, TestNonRetriableBelongsToItsOwnGroupAndLIF
   expected.push_back(std::make_pair(second_submitted, should_not_retry_));
 
   auto worker_to_kill_and_should_retry_ =
-      worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
+      worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot()).front();
 
   auto worker_to_kill = worker_to_kill_and_should_retry_.first;
   bool retry = worker_to_kill_and_should_retry_.second;
@@ -163,7 +162,7 @@ TEST_F(WorkerKillingGroupByOwnerTest, TestGroupSortedByGroupSizeThenFirstSubmitt
 
   for (const auto &entry : expected) {
     auto worker_to_kill_and_should_retry_ =
-        worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
+        worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot()).front();
     auto worker_to_kill = worker_to_kill_and_should_retry_.first;
     bool retry = worker_to_kill_and_should_retry_.second;
     ASSERT_EQ(worker_to_kill->WorkerId(), entry.first->WorkerId());
@@ -192,7 +191,7 @@ TEST_F(WorkerKillingGroupByOwnerTest, TestGroupSortedByRetriableLifo) {
 
   for (const auto &entry : expected) {
     auto worker_to_kill_and_should_retry_ =
-        worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
+        worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot()).front();
     auto worker_to_kill = worker_to_kill_and_should_retry_.first;
     bool retry = worker_to_kill_and_should_retry_.second;
     ASSERT_EQ(worker_to_kill->WorkerId(), entry.first->WorkerId());
@@ -218,7 +217,7 @@ TEST_F(WorkerKillingGroupByOwnerTest,
 
   for (const auto &entry : expected) {
     auto worker_to_kill_and_should_retry_ =
-        worker_killing_policy_.SelectWorkerToKill(workers, MemorySnapshot());
+        worker_killing_policy_.SelectWorkersToKill(workers, MemorySnapshot()).front();
     auto worker_to_kill = worker_to_kill_and_should_retry_.first;
     bool retry = worker_to_kill_and_should_retry_.second;
     ASSERT_EQ(worker_to_kill->WorkerId(), entry.first->WorkerId());
